layer4_alignment/requirements.txt
# Layer 4 Alignment Pipeline Dependencies
# ========================================

# Core
pydantic>=2.5.0
pyyaml>=6.0
aiosqlite>=0.19.0

# Alignment Methods
sentence-transformers>=2.2.0    # Vector embeddings (SBERT)
scikit-learn>=1.3.0             # TF-IDF, utilities

# LLM Providers (optional - enable in config)
google-generativeai>=0.3.0      # Gemini API
# openai>=1.6.0                 # OpenAI API (uncomment if needed)

# Utilities  
requests>=2.31.0                # Wikidata API
tqdm>=4.66.0                    # Progress bars
numpy>=1.24.0                   # Numerical operations

# Export
pandas>=2.0.0                   # CSV export, data manipulation
